added mutex with lock and unlock to type DriveFile

+++ lib/driveFile.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"sync"
 	"time"
 )
 
@@ -17,6 +18,7 @@ type DriveFile struct {
 	Modified time.Time
 	Created  time.Time
 	Root     bool
+	*sync.Mutex
 }
 
 // Attr returns the file attributes
@@ -30,6 +32,8 @@ func (d DriveFile) Attr() fuse.Attr {
 
 // ReadAll reads an entire file from google drive and returns the resulting bytes
 func (d DriveFile) ReadAll(intr fs.Intr) ([]byte, fuse.Error) {
+	d.Lock()
+	defer d.Unlock()
 	byteChan := make(chan *[]byte)
 	errChan := make(chan error)
 	defer func() {
@@ -66,6 +70,7 @@ func (d DriveFile) ReadAll(intr fs.Intr) ([]byte, fuse.Error) {
 
 // Setattr sets file attributs
 func (d *DriveFile) Setattr(req *fuse.SetattrRequest, resp *fuse.SetattrResponse, intr fs.Intr) fuse.Error {
+
 	valid := req.Valid
 	if valid.Size() {
 		return nil
@@ -75,6 +80,8 @@ func (d *DriveFile) Setattr(req *fuse.SetattrRequest, resp *fuse.SetattrResponse
 
 // Write writes bytes to a tmp file, which are then synced by FSync
 func (d *DriveFile) Write(req *fuse.WriteRequest, resp *fuse.WriteResponse, intr fs.Intr) fuse.Error {
+	d.Lock()
+	defer d.Unlock()
 	// check if d already has a tmp file
 	// If d does not have a tmp file, create one and write to it
 	size, err := d.TmpFile.Write(req.Data)
@@ -89,6 +96,8 @@ func (d *DriveFile) Write(req *fuse.WriteRequest, resp *fuse.WriteResponse, intr
 
 // Open a file or directory
 func (d *DriveFile) Open(req *fuse.OpenRequest, resp *fuse.OpenResponse, intr fs.Intr) (fs.Handle, fuse.Error) {
+	d.Lock()
+	defer d.Unlock()
 	d.TmpFile.Close()
 	f, err := os.Create("/tmp/drivefs-" + d.File.Id)
 	if err != nil {
@@ -102,7 +111,8 @@ func (d *DriveFile) Open(req *fuse.OpenRequest, resp *fuse.OpenResponse, intr fs
 
 // Flush a file tmp to google drive
 func (d *DriveFile) Flush(req *fuse.FlushRequest, intr fs.Intr) fuse.Error {
-	log.Println("Flushing: ", d.TmpFile.Name())
+	d.Lock()
+	defer d.Unlock()
 	// sync file to disk
 	err := d.TmpFile.Sync()
 	if err != nil {
@@ -132,9 +142,11 @@ func (d *DriveFile) Flush(req *fuse.FlushRequest, intr fs.Intr) fuse.Error {
 	// if all goes well, set new file and return
 	case f := <-fileChan:
 		d.File = f
-		return os.Remove(d.TmpFile.Name())
+		if d.TmpFile != nil && exists(d.TmpFile.Name()) {
+			return os.Remove(d.TmpFile.Name())
+		}
+		return nil
 	case err := <-errChan:
-		os.Remove(d.TmpFile.Name())
 		return err
 	// catch interupt
 	case <-intr:
@@ -151,4 +163,14 @@ func exists(fileName string) bool {
 		log.Println(a)
 		return true
 	}
+}
+
+// Lock locks a *DriveFile's mutex
+func (d *DriveFile) Lock() {
+	d.Mutex.Lock()
+}
+
+// Unlock unlocks a *DriveFile's mutex
+func (d *DriveFile) Unlock() {
+	d.Mutex.Unlock()
 }
